Refactor: DTO 경로 수정(response, request 폴더 분리)
