+++ src/main/java/Team_Mute/back_end/domain/reservation_admin/service/ReservationAdminService.java
@@ -10,11 +10,11 @@ import Team_Mute.back_end.domain.member.repository.UserRepository;
 import Team_Mute.back_end.domain.previsit.entity.PrevisitReservation;
 import Team_Mute.back_end.domain.reservation.entity.Reservation;
 import Team_Mute.back_end.domain.reservation.entity.ReservationStatus;
-import Team_Mute.back_end.domain.reservation_admin.dto.ApproveResponseDto;
-import Team_Mute.back_end.domain.reservation_admin.dto.PrevisitItemResponseDto;
-import Team_Mute.back_end.domain.reservation_admin.dto.RejectRequestDto;
-import Team_Mute.back_end.domain.reservation_admin.dto.RejectResponseDto;
-import Team_Mute.back_end.domain.reservation_admin.dto.ReservationListResponseDto;
+import Team_Mute.back_end.domain.reservation_admin.dto.request.RejectRequestDto;
+import Team_Mute.back_end.domain.reservation_admin.dto.response.ApproveResponseDto;
+import Team_Mute.back_end.domain.reservation_admin.dto.response.PrevisitItemResponseDto;
+import Team_Mute.back_end.domain.reservation_admin.dto.response.RejectResponseDto;
+import Team_Mute.back_end.domain.reservation_admin.dto.response.ReservationListResponseDto;
 import Team_Mute.back_end.domain.reservation_admin.entity.ReservationLog;
 import Team_Mute.back_end.domain.reservation_admin.repository.AdminPrevisitReservationRepository;
 import Team_Mute.back_end.domain.reservation_admin.repository.AdminReservationRepository;
